refactored the classes further
